+++ modules/session/classes/Session/ORM.php
@@ -29,33 +29,21 @@ namespace zesk;
  * @author kent
  */
 class Session_ORM extends ORM implements Interface_Session {
-	
+
 	/**
 	 * Original session data (to see if things change)
 	 *
 	 * @var array
 	 */
 	private $original = array();
-	
+
 	/**
 	 * Something changed?
 	 *
 	 * @var boolean
 	 */
 	private $changed = false;
-	
-	/**
-	 *
-	 * @return Object
-	 */
-	function really_store() {
-		if ($this->member_is_empty("cookie")) {
-			$this->delete();
-		} else {
-			return parent::store();
-		}
-	}
-	
+
 	/**
 	 *
 	 * {@inheritdoc}
@@ -127,21 +115,10 @@ class Session_ORM extends ORM implements Interface_Session {
 			$this->set_member("ip", "127.0.0.1");
 		}
 	}
-	function store() {
-		$session = $this->application->session(false);
-		if (!$session instanceof self || $this->id() !== $session->id()) {
-			$result = parent::store();
-			if ($result) {
-				$this->changed = false;
-			}
-			return $result;
-		}
-		return $this;
-	}
 	public function cookie_expire() {
 		return to_integer($this->option_path("cookie.expire"), 604800);
 	}
-	
+
 	/**
 	 * Set Session cookie
 	 *
@@ -151,7 +128,7 @@ class Session_ORM extends ORM implements Interface_Session {
 	private static function _generate_cookie() {
 		return md5("" . mt_rand(0, 999999999) . microtime());
 	}
-	
+
 	/**
 	 * Authenticate user at IP
 	 *
@@ -167,7 +144,7 @@ class Session_ORM extends ORM implements Interface_Session {
 		$this->set_member("expires", Timestamp::now()->add_unit($cookieExpire, Timestamp::UNIT_SECOND));
 		return $this->store();
 	}
-	
+
 	/**
 	 * Are we authenticated?
 	 *
@@ -176,7 +153,7 @@ class Session_ORM extends ORM implements Interface_Session {
 	public function authenticated() {
 		return $this->member_is_empty('User');
 	}
-	
+
 	/**
 	 * De-authenticate
 	 *
@@ -189,7 +166,7 @@ class Session_ORM extends ORM implements Interface_Session {
 		$this->set_member("user", null);
 		return $this->store();
 	}
-	
+
 	/**
 	 * Is this session expired?
 	 *
@@ -199,7 +176,7 @@ class Session_ORM extends ORM implements Interface_Session {
 	public function expires() {
 		return $this->member_timestamp("expires");
 	}
-	
+
 	/**
 	 * Logout expired, run hook
 	 */
@@ -232,7 +209,7 @@ class Session_ORM extends ORM implements Interface_Session {
 			->where($where)
 			->execute();
 	}
-	
+
 	/**
 	 *
 	 * @return Timestamp
@@ -245,7 +222,7 @@ class Session_ORM extends ORM implements Interface_Session {
 	private function cookie_name() {
 		return $this->option_path("cookie.name", "ZCOOKIE");
 	}
-	
+
 	/**
 	 *
 	 * {@inheritdoc}
@@ -275,9 +252,14 @@ class Session_ORM extends ORM implements Interface_Session {
 		$application->hooks->add(Response::class . "::headers", function (Response $response) use ($cookie_name, $cookie_value, $cookie_options) {
 			$response->cookie($cookie_name, $cookie_value, $cookie_options);
 		});
+		$session = $this;
+		$application->hooks->add(Hooks::hook_exit, function () use ($session) {
+			$session->store();
+		});
+
 		return $this;
 	}
-	
+
 	/**
 	 *
 	 * @return string
@@ -285,7 +267,7 @@ class Session_ORM extends ORM implements Interface_Session {
 	function hash() {
 		return $this->member("cookie");
 	}
-	
+
 	/**
 	 *
 	 * @param unknown $user
@@ -337,7 +319,7 @@ class Session_ORM extends ORM implements Interface_Session {
 		$this->set_master();
 		return $this;
 	}
-	
+
 	/**
 	 * Count all other sessions seen within the seconds window provided
 	 *
@@ -352,7 +334,7 @@ class Session_ORM extends ORM implements Interface_Session {
 			->where($where)
 			->one_integer("X");
 	}
-	
+
 	/*
 	 * Get/Set session valuesfrom Object
 	 *
@@ -362,6 +344,19 @@ class Session_ORM extends ORM implements Interface_Session {
 			$this->members['data'] = array();
 		}
 		$this->original = $this->members['data'];
+	}
+
+	/**
+	 *
+	 * @return Object
+	 */
+	function store() {
+		if ($this->member_is_empty("cookie")) {
+			$this->delete();
+			return $this;
+		} else {
+			return parent::store();
+		}
 	}
 	public function user_id() {
 		return $this->member_integer("user");
@@ -382,7 +377,7 @@ class Session_ORM extends ORM implements Interface_Session {
 		}
 		return $default;
 	}
-	
+
 	/**
 	 * Session variables are special
 	 *
@@ -391,7 +386,7 @@ class Session_ORM extends ORM implements Interface_Session {
 	public function __get($name) {
 		return avalue($this->members['data'], $name);
 	}
-	
+
 	/**
 	 * (non-PHPdoc)
 	 *
@@ -414,7 +409,7 @@ class Session_ORM extends ORM implements Interface_Session {
 	public function changed($members = null) {
 		return $this->changed;
 	}
-	
+
 	/**
 	 * Retrieve some of the values
 	 *
@@ -434,7 +429,7 @@ class Session_ORM extends ORM implements Interface_Session {
 	public function variables() {
 		return $this->members['data'];
 	}
-	
+
 	/**
 	 *
 	 * @return self
@@ -442,21 +437,7 @@ class Session_ORM extends ORM implements Interface_Session {
 	public function found_session() {
 		return $this;
 	}
-	
-	/**
-	 * Really save
-	 */
-	public static function save(Application $application) {
-		try {
-			$session = $application->session(false);
-			if ($session instanceof self) {
-				$session->really_store();
-			}
-		} catch (Database_Exception_Connect $e) {
-		} catch (Database_Exception_Table_NotFound $e) {
-		}
-	}
-	
+
 	/**
 	 *
 	 * @return array